Document use of OSH and create $(SH) predefine.
MANUAL.md has a new subsection on Running OSS commands.
default.c has the beginnings of predefines for GUARDIAN. This should not
be expanded at this time because the GNUMake 4.3 version reworks the
structure of default.c.
.gitignore modified to add ZLDAP* to avoid committing junk objects.

Fixes #50

CLA: CLA on file for ITUGLIB use of this fix partially applied from GMAKEN fork.

Signed-off-by: Randall S. Becker <rsbecker@nexbridge.com>