Add path type resolution in tandem_ext.c for simulating stat.
This includes handling a ./ prefix and . alone passed to stat(). Processing
for patterns is done by converting Guardian into OSS and reading the OSS
directory to perform pattern matching. The matched names are then converted
back to Guardian for use by the makefile. Conversion methods are added to
tandem_ext.c.

stat() is made aware of this and handles OSS names. Directories and sub-volumes
passed to stat() are simulated and do not contain actual information.

The manual has also been updated to document $(wildcard) behaviour.

Unit test t0001 added to test conversion methods.

Remove RSC and SCCS implicit rules that are incompatible with this change.

Fixes: #102

Signed-off-by: Randall S. Becker <randall.becker@nexbridge.ca>